backend/requirements.txt
amqp==5.0.2
bcrypt==3.2.0
billiard==3.6.3.0
celery==5.0.2
certifi==2020.11.8
cffi==1.14.5
chardet==3.0.4
click==7.1.2
click-didyoumean==0.0.3
click-repl==0.1.6
ecdsa==0.14.1
Flask==1.1.2
Flask-Cors==3.0.9
Flask-JWT-Extended==3.25.0
idna==2.10
itsdangerous==1.1.0
Jinja2==2.11.2
kombu==5.0.2
MarkupSafe==1.1.1
passlib==1.7.4
prompt-toolkit==3.0.8
psycopg2-binary==2.8.6
pyasn1==0.4.8
pycparser==2.20
PyJWT==1.7.1
python-jose==3.2.0
pytz==2020.4
redis==3.5.3
requests==2.25.0
rsa==4.7.2
six==1.15.0
SQLAlchemy==1.3.20
urllib3==1.26.2
vine==5.0.0
wcwidth==0.2.5
Werkzeug==1.0.1
wincertstore==0.2
